+++ app/scripts/HiGlassComponent.jsx
@@ -2082,7 +2082,7 @@ class HiGlassComponent extends React.Component {
       };
 
       this.addCallbacks(toView, newTrack);
-      this.handleTrackAdded(toView, newTrack, position, hostTrack);
+      this.handleTrackAdded(toView, newTrack, position, null, hostTrack);
     }
     this.setState({
       chooseTrackHandler: null,
@@ -2155,9 +2155,10 @@ class HiGlassComponent extends React.Component {
    *
    * We just need to close the menu here.
    */
-  handleTrackPositionChosen(viewUid, position) {
+  handleTrackPositionChosen(viewUid, position, extent) {
     this.setState({
       addTrackPosition: position,
+      addTrackExtent: extent,
       addTrackPositionView: viewUid,
     });
   }
@@ -2708,6 +2709,7 @@ class HiGlassComponent extends React.Component {
       // we've already added the track, remove the add track dialog
       this.setState({
         addTrackPosition: null,
+        addTrackExtent: null,
       });
     }
   }
@@ -2720,11 +2722,11 @@ class HiGlassComponent extends React.Component {
    * @param position: The position the track is being added to
    * @param host: If this track is being added to another track
    */
-  handleTracksAdded(viewId, newTracks, position, host) {
+  handleTracksAdded(viewId, newTracks, position, extent, host) {
     this.storeTrackSizes(viewId);
 
     for (const newTrack of newTracks) {
-      this.handleTrackAdded(viewId, newTrack, position, host);
+      this.handleTrackAdded(viewId, newTrack, position, extent, host);
     }
   }
 
@@ -2793,7 +2795,7 @@ class HiGlassComponent extends React.Component {
    * @returns {Object}: A trackConfig (\{ uid: "", width: x \})
    *  describing this track
    */
-  handleTrackAdded(viewId, newTrack, position, host = null) {
+  handleTrackAdded(viewId, newTrack, position, extent = null, host = null) {
     this.addDefaultTrackOptions(newTrack);
 
     // make sure the new track has a uid
@@ -2810,6 +2812,7 @@ class HiGlassComponent extends React.Component {
       // we've already added the track, remove the add track dialog
       this.setState({
         addTrackPosition: null,
+        addTrackExtent: null,
       });
     }
 
@@ -2820,7 +2823,9 @@ class HiGlassComponent extends React.Component {
     }
 
     newTrack.position = position;
-
+    if (extent) {
+      newTrack.options.extent = extent;
+    }
     const trackInfo = this.getTrackInfo(newTrack.type);
 
     newTrack.width =
@@ -3883,7 +3888,7 @@ class HiGlassComponent extends React.Component {
           if (
             otherTrack.type === 'heatmap' &&
             otherTrack.uid !== track.uid &&
-            otherTrack.options.extent !== 'lower-left'
+            otherTrack.options.extent === 'full'
           ) {
             // Automatically change the extent of the other track to
             // `lower-left``
@@ -3928,7 +3933,7 @@ class HiGlassComponent extends React.Component {
             if (
               otherTrack.type === 'heatmap' &&
               otherTrack.uid !== track.uid &&
-              otherTrack.options.extent !== 'upper-right'
+              otherTrack.options.extent === 'full'
             ) {
               // Automatically change the extent of the other track to
               // `upper-right``
@@ -5024,6 +5029,11 @@ class HiGlassComponent extends React.Component {
               this.tiledPlots[view.uid] = c;
             }}
             // Custom props
+            addTrackExtent={
+              this.state.addTrackPositionView === view.uid
+                ? this.state.addTrackExtent
+                : null
+            }
             addTrackPosition={
               this.state.addTrackPositionView === view.uid
                 ? this.state.addTrackPosition
@@ -5084,8 +5094,14 @@ class HiGlassComponent extends React.Component {
               this.handleTrackOptionsChanged(view.uid, trackId, options)
             }
             onTrackPositionChosen={this.handleTrackPositionChosen.bind(this)}
-            onTracksAdded={(newTracks, position, host) =>
-              this.handleTracksAdded(view.uid, newTracks, position, host)
+            onTracksAdded={(newTracks, position, extent, host) =>
+              this.handleTracksAdded(
+                view.uid,
+                newTracks,
+                position,
+                extent,
+                host,
+              )
             }
             onUnlockValueScale={(uid) =>
               this.handleUnlockValueScale(view.uid, uid)
@@ -5242,8 +5258,8 @@ class HiGlassComponent extends React.Component {
               onTogglePositionSearchBox={this.handleTogglePositionSearchBox.bind(
                 this,
               )}
-              onTrackPositionChosen={(position) =>
-                this.handleTrackPositionChosen(view.uid, position)
+              onTrackPositionChosen={(position, extent) =>
+                this.handleTrackPositionChosen(view.uid, position, extent)
               }
               onUnlockLocation={(uid) =>
                 this.handleUnlock(uid, this.locationLocks)