+++ app/scripts/TiledPlot.jsx
@@ -112,6 +112,7 @@ export class TiledPlot extends React.Component {
 
       tracks,
       init: false,
+      addTrackExtent: null,
       addTrackPosition: null,
       customDialog: null,
       mouseOverOverlayUid: null,
@@ -771,6 +772,7 @@ export class TiledPlot extends React.Component {
     if (this.state.addTrackPosition || this.props.addTrackPosition) {
       this.props.modal.open(
         <AddTrackDialog
+          extent={this.state.addTrackExtent || this.props.addTrackExtent}
           host={this.state.addTrackHost}
           onCancel={this.handleNoTrackAddedBound}
           onTracksChosen={this.handleTracksAddedBound}
@@ -1020,6 +1022,7 @@ export class TiledPlot extends React.Component {
     this.props.onNoTrackAdded();
 
     this.setState({
+      addTrackExtent: null,
       addTrackPosition: null,
       addTrackHost: null,
     });
@@ -1091,8 +1094,9 @@ export class TiledPlot extends React.Component {
     this.handleAddTrack(orientation);
   }
 
-  handleAddTrack(position) {
+  handleAddTrack(position, extent) {
     this.setState({
+      addTrackExtent: extent,
       addTrackPosition: position,
       addTrackHost: null,
     });
@@ -1166,7 +1170,7 @@ export class TiledPlot extends React.Component {
     this.props.onChangeTrackData(uid, newData);
   }
 
-  handleTracksAdded(newTracks, position, host) {
+  handleTracksAdded(newTracks, position, extent, host) {
     /**
      * Arguments
      * ---------
@@ -1194,9 +1198,10 @@ export class TiledPlot extends React.Component {
 
     // if host is defined, then we're adding a new series
     // further down the chain a combined track will be created
-    this.props.onTracksAdded(newTracks, position, host);
+    this.props.onTracksAdded(newTracks, position, extent, host);
 
     this.setState({
+      addTrackExtent: null,
       addTrackPosition: null,
       addTrackHost: null,
     });
@@ -1837,6 +1842,7 @@ export class TiledPlot extends React.Component {
       overlays: props.overlays,
       viewOptions: props.viewOptions,
       uid: props.uid,
+      addTrackExtent: props.addTrackExtent,
       addTrackPosition: props.addTrackPosition,
       editable: props.editable,
       marginTop: props.marginTop,
@@ -2061,7 +2067,12 @@ export class TiledPlot extends React.Component {
             // because "whole" tracks are always drawn on top of each other,
             // the notion of Series is unnecessary and so 'host' is null
             onAddTrack={(newTrack) => {
-              this.props.onTracksAdded([newTrack], newTrack.position, null);
+              this.props.onTracksAdded(
+                [newTrack],
+                newTrack.position,
+                null,
+                null,
+              );
               this.handleCloseContextMenu();
             }}
             onChangeTrackData={this.handleChangeTrackDataBound}
@@ -2830,6 +2841,7 @@ TiledPlot.defaultProps = {
 };
 
 TiledPlot.propTypes = {
+  addTrackExtent: PropTypes.string,
   addTrackPosition: PropTypes.string,
   canvasElement: PropTypes.object,
   chooseTrackHandler: PropTypes.func,