+++ test/utils/create-element-and-api.js
@@ -1,8 +1,15 @@
-// @ts-nocheck
-import { viewer } from '../../app/scripts/hglib';
-
+import { test } from 'vitest';
+import * as hglib from '../../app/scripts/hglib';
 import createDiv from './create-div';
 
+/**
+ * @param {Record<string, unknown>} viewConfig
+ * @param {Record<string, unknown>} options
+ * @param {number} [width]
+ * @param {number} [height]
+ * @param {boolean} [scrollable]
+ * @return {Promise<[HTMLDivElement, hglib.HiGlassApi]>}
+ */
 export default async function createElementAndApi(
   viewConfig,
   options,
@@ -22,7 +29,24 @@ export default async function createElementAndApi(
     }`,
   );
 
-  const api = await viewer(div, viewConfig, options);
+  const api = await hglib.viewer(div, viewConfig, options);
 
   return [div, api];
+}
+
+/**
+ * @param {Record<string, unknown>} viewconf
+ * @returns {import("vitest").TestAPI<{api: hglib.HiGlassApi }>}
+ */
+export function createHiGlassTestApi(viewconf) {
+  return test.extend({
+    // biome-ignore lint/correctness/noEmptyPattern: Needed for vitest
+    api: async ({}, use) => {
+      const [div, api] = await createElementAndApi(viewconf, { bounded: true });
+      document.body.append(div);
+      await use(api);
+      api.destroy();
+      document.body.replaceChildren();
+    },
+  });
 }